refactor(providers): extract provider selection policy from ProviderRegistry
ProviderRegistry mixed registration, strategy lifecycle, and ~430 LOC of
provider selection/load-balancing. Extract the selection policy to a
ProviderSelectionService implementing the existing ProviderSelectionPort
(depending only on abstractions, so no circular dependency — the
prior DO NOT MOVE comment was over-cautious). ProviderRegistry drops to
registration + strategy lifecycle.

- New: src/orb/infrastructure/services/provider_selection_service.py
  holds select_provider_for_template, select_active_provider, and all
  _select_by_*, _apply_load_balancing_strategy, _weighted_round_robin,
  _health_based_selection, _find_compatible_providers, _provider_supports_api

- ProviderRegistry: thin delegators for select_* methods, lazy-constructs
  ProviderSelectionService on first call (bootstrap wires it eagerly)

- ProviderRegistryPort: two new abstract methods — get_strategy (read-only
  cache access needed by _provider_supports_api) and get_fallback_strategy
  (needed by _select_default_provider)

- bootstrap/provider_services.py: wires ProviderSelectionService eagerly
  when constructing the registry singleton so config_port is shared

- Tests: update __new__-based helpers to set _selection_service = None;
  source-inspection tests now reference ProviderSelectionService

Before/after LOC: provider_registry.py 1062 → 710 (-33%)
